+++ device/stub/instamsg/device_time.c
@@ -8,7 +8,7 @@
 /*
  * This method returns the minimum-delay achievable via this device.
  */
-unsigned int getMinimumDelayPossibleInMicroSeconds()
+unsigned long getMinimumDelayPossibleInMicroSeconds()
 {
     return 0;
 }
